+++ packages/lib-shared/js/internal/NativeWrapper.ts
@@ -92,7 +92,7 @@ class DebugThisCall implements ThisCallTrampoline {
             }
             const r = await NativeModulesProvider.PowerAuth.callNative<T>(name, [instanceId, ...args])
             if (this.traceCall) {
-                console.log(` ret ${msg} => ${JSON.stringify(r)}`)
+                console.log(` ret ${msg} => ${prettyResult(name, r)}`)
             }
             return r
         } catch (e) {
@@ -296,10 +296,14 @@ function prettyArgs(fname: string, args: any[]): string {
         case 'unsafeChangePassword': 
             sanitizedArgs[1] = sanitizedArgs[2] = '***'
             break
+        case 'beginPasswordChange':
         case 'validatePassword':
         case 'addBiometryFactor':
         case 'startProtocolUpgrade':
             sanitizedArgs[1] = '***'
+            break
+        case 'finishPasswordChange':
+            sanitizedArgs[1] = sanitizedArgs[2] = '***'
             break
         case 'configure': {
             const clientConfiguration = sanitizedArgs[2]
@@ -335,4 +339,8 @@ function prettyArgs(fname: string, args: any[]): string {
         return value
     })
     return v.slice(1, v.length - 1)
+}
+
+function prettyResult(fname: string, result: any): string {
+    return fname === 'beginPasswordChange' ? '"***"' : JSON.stringify(result)
 }